requirements.txt
langchain
langchain-groq
python-dotenv
langchain_community
streamlit
